+++ include/filter_test/residuals/constant_residual.h
@@ -8,6 +8,7 @@
 #ifndef INCLUDE_FILTER_TEST_RESIDUALS_CONSTANT_RESIDUAL_H_
 #define INCLUDE_FILTER_TEST_RESIDUALS_CONSTANT_RESIDUAL_H_
 
+#include "filter_test/block_impl.h"
 #include "filter_test/residual.h"
 
 namespace tsif {
