+++ include/filter_test/residuals/constant_velocity_residual.h
@@ -8,6 +8,9 @@
 #ifndef INCLUDE_FILTER_TEST_RESIDUALS_CONSTANT_VELOCITY_RESIDUAL_H_
 #define INCLUDE_FILTER_TEST_RESIDUALS_CONSTANT_VELOCITY_RESIDUAL_H_
 
+#include "filter_test/residual.h"
+#include "filter_test/block_impl.h"
+
 namespace tsif {
 
 // This residual implements the following set of equations
